Refactor code for SARS-CoV-2 test simulation
This commit refactors the code for the SARS-CoV-2 test simulation. The changes include setting up the lists for random.choice to generate infected and uninfected people, as well as positive and negative test outcomes. The code also includes a simulation for counting the number of correct and incorrect test results.